Phase 1: extract platform/ios-chrome.js (+ migrate tabs to state)
Extracts the iOS native chrome + print cluster (~280 lines) into
platform/ios-chrome.js: the action bar / sheets sync (syncIOSChrome), the
window.webkit message bridges (open file, bundled sample, native print), the
window.setIOSLayoutMode bridge, sheet show/hide helpers, and the print pipeline
(performPrint + buildPrintableDocument, which is entangled with iOS native
print). It imports state/platform/escapeHtml and looks up its DOM by id.

To enable it, migrated the final shared field `tabs` into core/state.js
(syncIOSChrome reads state.tabs.length) — done with a code-only rewrite so the
user-facing "…tabs reached" string and a comment are untouched. Removed three
now-dead DOM consts from main.js (iosSampleSection, iosActionBar, searchCount).

Pure refactor. main.js ~2,225 -> ~1,943 lines (from 2,814 at the split's start).
Build + lint clean; 297 tests pass.
