+++ tests/unit/desktopRenderer.test.js
@@ -53,8 +53,8 @@ describe('Desktop renderer integration', () => {
     createTab('one.md', '# One again', '/tmp/one.md');
     expect(window.specdown.watchFile).toHaveBeenCalledTimes(1);
 
-    const firstTabId = tabs[0].id;
-    const secondTabId = tabs[1].id;
+    const firstTabId = state.tabs[0].id;
+    const secondTabId = state.tabs[1].id;
 
     await closeTab(firstTabId);
     expect(window.specdown.unwatchFile).not.toHaveBeenCalled();
@@ -73,8 +73,8 @@ describe('Desktop renderer integration', () => {
     createTab('b.md', '# B', '/tmp/b.md');
 
     // Tab b is now active; deliver a file-changed event for tab a.
-    const backgroundTab = tabs.find(t => t.filePath === '/tmp/a.md');
-    const activeTab = tabs.find(t => t.filePath === '/tmp/b.md');
+    const backgroundTab = state.tabs.find(t => t.filePath === '/tmp/a.md');
+    const activeTab = state.tabs.find(t => t.filePath === '/tmp/b.md');
     expect(state.activeTabId).toBe(activeTab.id);
 
     await fileChangedCallback({
